collatz.cs
This is a very simple C# program that takes a command line argument, and then tests that value against the Collatz Conjecture. Basically, if a number is even, divide it by 2. If the number is odd, multiply by 3 and add 1. If you continue this pattern, you should eventually get to 1.